Split Roblox-specific functionality into dedicated interface
