+++ src/LuauExt.cpp
@@ -1,3 +1,6 @@
+#include <utility>
+
+#include "Platform/LSPPlatform.hpp"
 #include "Luau/BuiltinDefinitions.h"
 #include "Luau/ToString.h"
 #include "Luau/Transpiler.h"
@@ -60,391 +63,7 @@ std::optional<std::string> getTypeName(Luau::TypeId typeId)
         return name;
 }
 
-// Retrieves the corresponding Luau type for a Sourcemap node
-// If it does not yet exist, the type is produced
-Luau::TypeId getSourcemapType(const Luau::GlobalTypes& globals, Luau::TypeArena& arena, const SourceNodePtr& node)
-{
-    // Gets the type corresponding to the sourcemap node if it exists
-    // Make sure to use the correct ty version (base typeChecker vs autocomplete typeChecker)
-    if (node->tys.find(&globals) != node->tys.end())
-        return node->tys.at(&globals);
-
-    Luau::LazyType ltv(
-        [&globals, &arena, node](Luau::LazyType& ltv) -> void
-        {
-            // Check if the LTV already has an unwrapped type
-            if (ltv.unwrapped.load())
-                return;
-
-            // Handle if the node is no longer valid
-            if (!node)
-            {
-                ltv.unwrapped = globals.builtinTypes->anyType;
-                return;
-            }
-
-            auto instanceTy = globals.globalScope->lookupType("Instance");
-            if (!instanceTy)
-            {
-                ltv.unwrapped = globals.builtinTypes->anyType;
-                return;
-            }
-
-            // Look up the base class instance
-            Luau::TypeId baseTypeId = getTypeIdForClass(globals.globalScope, node->className).value_or(nullptr);
-            if (!baseTypeId)
-            {
-                ltv.unwrapped = globals.builtinTypes->anyType;
-                return;
-            }
-
-            // Point the metatable to the metatable of "Instance" so that we allow equality
-            std::optional<Luau::TypeId> instanceMetaIdentity;
-            if (auto* ctv = Luau::get<Luau::ClassType>(instanceTy->type))
-                instanceMetaIdentity = ctv->metatable;
-
-            // Create the ClassType representing the instance
-            std::string typeName = getTypeName(baseTypeId).value_or(node->name);
-            Luau::ClassType baseInstanceCtv{typeName, {}, baseTypeId, instanceMetaIdentity, {}, {}, "@roblox"};
-            auto typeId = arena.addType(std::move(baseInstanceCtv));
-
-            // Attach Parent and Children info
-            // Get the mutable version of the type var
-            if (auto* ctv = Luau::getMutable<Luau::ClassType>(typeId))
-            {
-                if (auto parentNode = node->parent.lock())
-                    ctv->props["Parent"] = Luau::makeProperty(getSourcemapType(globals, arena, parentNode));
-
-                // Add children as properties
-                for (const auto& child : node->children)
-                    ctv->props[child->name] = Luau::makeProperty(getSourcemapType(globals, arena, child));
-
-                // Add FindFirstAncestor and FindFirstChild
-                if (auto instanceType = getTypeIdForClass(globals.globalScope, "Instance"))
-                {
-                    auto findFirstAncestorFunction = Luau::makeFunction(arena, typeId, {globals.builtinTypes->stringType}, {"name"}, {*instanceType});
-
-                    Luau::attachMagicFunction(findFirstAncestorFunction,
-                        [&arena, &globals, node](Luau::TypeChecker& typeChecker, const Luau::ScopePtr& scope, const Luau::AstExprCall& expr,
-                            const Luau::WithPredicate<Luau::TypePackId>& withPredicate) -> std::optional<Luau::WithPredicate<Luau::TypePackId>>
-                        {
-                            if (expr.args.size < 1)
-                                return std::nullopt;
-
-                            auto str = expr.args.data[0]->as<Luau::AstExprConstantString>();
-                            if (!str)
-                                return std::nullopt;
-
-                            // This is a O(n) search, not great!
-                            if (auto ancestor = node->findAncestor(std::string(str->value.data, str->value.size)))
-                            {
-                                return Luau::WithPredicate<Luau::TypePackId>{arena.addTypePack({getSourcemapType(globals, arena, *ancestor)})};
-                            }
-
-                            return std::nullopt;
-                        });
-                    ctv->props["FindFirstAncestor"] = Luau::makeProperty(findFirstAncestorFunction, "@roblox/globaltype/Instance.FindFirstAncestor");
-
-                    auto findFirstChildFunction = Luau::makeFunction(arena, typeId, {globals.builtinTypes->stringType}, {"name"}, {*instanceType});
-                    Luau::attachMagicFunction(findFirstChildFunction,
-                        [node, &arena, &globals](Luau::TypeChecker& typeChecker, const Luau::ScopePtr& scope, const Luau::AstExprCall& expr,
-                            const Luau::WithPredicate<Luau::TypePackId>& withPredicate) -> std::optional<Luau::WithPredicate<Luau::TypePackId>>
-                        {
-                            if (expr.args.size < 1)
-                                return std::nullopt;
-
-                            auto str = expr.args.data[0]->as<Luau::AstExprConstantString>();
-                            if (!str)
-                                return std::nullopt;
-
-                            if (auto child = node->findChild(std::string(str->value.data, str->value.size)))
-                                return Luau::WithPredicate<Luau::TypePackId>{arena.addTypePack({getSourcemapType(globals, arena, *child)})};
-
-                            return std::nullopt;
-                        });
-                    ctv->props["FindFirstChild"] = Luau::makeProperty(findFirstChildFunction, "@roblox/globaltype/Instance.FindFirstChild");
-                }
-            }
-
-            ltv.unwrapped = typeId;
-            return;
-        });
-    auto ty = arena.addType(std::move(ltv));
-    node->tys.insert_or_assign(&globals, ty);
-
-    return ty;
-}
-
-// Magic function for `Instance:IsA("ClassName")` predicate
-std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionInstanceIsA(Luau::TypeChecker& typeChecker, const Luau::ScopePtr& scope,
-    const Luau::AstExprCall& expr, const Luau::WithPredicate<Luau::TypePackId>& withPredicate)
-{
-    if (expr.args.size != 1)
-        return std::nullopt;
-
-    auto index = expr.func->as<Luau::AstExprIndexName>();
-    auto str = expr.args.data[0]->as<Luau::AstExprConstantString>();
-    if (!index || !str)
-        return std::nullopt;
-
-    std::optional<Luau::LValue> lvalue = tryGetLValue(*index->expr);
-    if (!lvalue)
-        return std::nullopt;
-
-    std::string className(str->value.data, str->value.size);
-    std::optional<Luau::TypeFun> tfun = typeChecker.globalScope->lookupType(className);
-    if (!tfun || !tfun->typeParams.empty() || !tfun->typePackParams.empty())
-    {
-        typeChecker.reportError(Luau::TypeError{expr.args.data[0]->location, Luau::UnknownSymbol{className, Luau::UnknownSymbol::Type}});
-        return std::nullopt;
-    }
-
-    auto type = Luau::follow(tfun->type);
-
-    Luau::TypeArena& arena = typeChecker.currentModule->internalTypes;
-    Luau::TypePackId booleanPack = arena.addTypePack({typeChecker.booleanType});
-    return Luau::WithPredicate<Luau::TypePackId>{booleanPack, {Luau::IsAPredicate{std::move(*lvalue), expr.location, type}}};
-}
-
-// Magic function for `instance:Clone()`, so that we return the exact subclass that `instance` is, rather than just a generic Instance
-std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionInstanceClone(Luau::TypeChecker& typeChecker, const Luau::ScopePtr& scope,
-    const Luau::AstExprCall& expr, const Luau::WithPredicate<Luau::TypePackId>& withPredicate)
-{
-    auto index = expr.func->as<Luau::AstExprIndexName>();
-    if (!index)
-        return std::nullopt;
-
-    Luau::TypeArena& arena = typeChecker.currentModule->internalTypes;
-    auto instanceType = typeChecker.checkExpr(scope, *index->expr);
-    return Luau::WithPredicate<Luau::TypePackId>{arena.addTypePack({instanceType.type})};
-}
-
-// Magic function for `Instance:FindFirstChildWhichIsA("ClassName")` and friends
-std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionFindFirstXWhichIsA(Luau::TypeChecker& typeChecker, const Luau::ScopePtr& scope,
-    const Luau::AstExprCall& expr, const Luau::WithPredicate<Luau::TypePackId>& withPredicate)
-{
-    if (expr.args.size < 1)
-        return std::nullopt;
-
-    auto str = expr.args.data[0]->as<Luau::AstExprConstantString>();
-    if (!str)
-        return std::nullopt;
-
-    std::optional<Luau::TypeFun> tfun = scope->lookupType(std::string(str->value.data, str->value.size));
-    if (!tfun || !tfun->typeParams.empty() || !tfun->typePackParams.empty())
-        return std::nullopt;
-
-    auto type = Luau::follow(tfun->type);
-
-    Luau::TypeArena& arena = typeChecker.currentModule->internalTypes;
-    Luau::TypeId nillableClass = Luau::makeOption(typeChecker.builtinTypes, arena, type);
-    return Luau::WithPredicate<Luau::TypePackId>{arena.addTypePack({nillableClass})};
-}
-
-// Magic function for `EnumItem:IsA("EnumType")` predicate
-std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionEnumItemIsA(Luau::TypeChecker& typeChecker, const Luau::ScopePtr& scope,
-    const Luau::AstExprCall& expr, const Luau::WithPredicate<Luau::TypePackId>& withPredicate)
-{
-    if (expr.args.size != 1)
-        return std::nullopt;
-
-    auto index = expr.func->as<Luau::AstExprIndexName>();
-    auto str = expr.args.data[0]->as<Luau::AstExprConstantString>();
-    if (!index || !str)
-        return std::nullopt;
-
-    std::optional<Luau::LValue> lvalue = tryGetLValue(*index->expr);
-    if (!lvalue)
-        return std::nullopt;
-
-    std::string enumItem(str->value.data, str->value.size);
-    std::optional<Luau::TypeFun> tfun = scope->lookupImportedType("Enum", enumItem);
-    if (!tfun || !tfun->typeParams.empty() || !tfun->typePackParams.empty())
-    {
-        typeChecker.reportError(Luau::TypeError{expr.args.data[0]->location, Luau::UnknownSymbol{enumItem, Luau::UnknownSymbol::Type}});
-        return std::nullopt;
-    }
-
-    auto type = Luau::follow(tfun->type);
-
-    Luau::TypeArena& arena = typeChecker.currentModule->internalTypes;
-    Luau::TypePackId booleanPack = arena.addTypePack({typeChecker.booleanType});
-    return Luau::WithPredicate<Luau::TypePackId>{booleanPack, {Luau::IsAPredicate{std::move(*lvalue), expr.location, type}}};
-}
-
-// Magic function for `instance:GetPropertyChangedSignal()`, so that we can perform type checking on the provided property
-static std::optional<Luau::WithPredicate<Luau::TypePackId>> magicFunctionGetPropertyChangedSignal(Luau::TypeChecker& typeChecker,
-    const Luau::ScopePtr& scope, const Luau::AstExprCall& expr, const Luau::WithPredicate<Luau::TypePackId>& withPredicate)
-{
-    if (expr.args.size != 1)
-        return std::nullopt;
-
-    auto index = expr.func->as<Luau::AstExprIndexName>();
-    auto str = expr.args.data[0]->as<Luau::AstExprConstantString>();
-    if (!index || !str)
-        return std::nullopt;
-
-
-    auto instanceType = typeChecker.checkExpr(scope, *index->expr);
-    auto ctv = Luau::get<Luau::ClassType>(Luau::follow(instanceType.type));
-    if (!ctv)
-        return std::nullopt;
-
-    std::string property(str->value.data, str->value.size);
-    if (!Luau::lookupClassProp(ctv, property))
-    {
-        typeChecker.reportError(Luau::TypeError{expr.args.data[0]->location, Luau::UnknownProperty{instanceType.type, property}});
-        return std::nullopt;
-    }
-
-    return std::nullopt;
-}
-
-void addChildrenToCTV(const Luau::GlobalTypes& globals, Luau::TypeArena& arena, const Luau::TypeId& ty, const SourceNodePtr& node)
-{
-    if (auto* ctv = Luau::getMutable<Luau::ClassType>(ty))
-    {
-        // Clear out all the old registered children
-        for (auto it = ctv->props.begin(); it != ctv->props.end();)
-        {
-            if (hasTag(it->second, "@sourcemap-generated"))
-                it = ctv->props.erase(it);
-            else
-                ++it;
-        }
-
-
-        // Extend the props to include the children
-        for (const auto& child : node->children)
-        {
-            ctv->props[child->name] = Luau::Property{
-                getSourcemapType(globals, arena, child),
-                /* deprecated */ false,
-                /* deprecatedSuggestion */ {},
-                /* location */ std::nullopt,
-                /* tags */ {"@sourcemap-generated"},
-                /* documentationSymbol*/ std::nullopt,
-            };
-        }
-    }
-}
-
-// TODO: expressiveTypes is used because of a Luau issue where we can't cast a most specific Instance type (which we create here)
-// to another type. For the time being, we therefore make all our DataModel instance types marked as "any".
-// Remove this once Luau has improved
-void registerInstanceTypes(Luau::Frontend& frontend, const Luau::GlobalTypes& globals, Luau::TypeArena& arena,
-    const WorkspaceFileResolver& fileResolver, bool expressiveTypes)
-{
-    if (!fileResolver.rootSourceNode)
-        return;
-
-    // Create a type for the root source node
-    getSourcemapType(globals, arena, fileResolver.rootSourceNode);
-
-    // Modify sourcemap types
-    if (fileResolver.rootSourceNode->className == "DataModel")
-    {
-        // Mutate DataModel with its children
-        if (auto dataModelType = globals.globalScope->lookupType("DataModel"))
-            addChildrenToCTV(globals, arena, dataModelType->type, fileResolver.rootSourceNode);
-
-        // Mutate globally-registered Services to include children information (so it's available through :GetService)
-        for (const auto& service : fileResolver.rootSourceNode->children)
-        {
-            auto serviceName = service->className; // We know it must be a service of the same class name
-            if (auto serviceType = globals.globalScope->lookupType(serviceName))
-                addChildrenToCTV(globals, arena, serviceType->type, service);
-        }
-
-        // Add containers to player and copy over instances
-        // TODO: Player.Character should contain StarterCharacter instances
-        if (auto playerType = globals.globalScope->lookupType("Player"))
-        {
-            if (auto* ctv = Luau::getMutable<Luau::ClassType>(playerType->type))
-            {
-                // Player.Backpack should be defined
-                if (auto backpackType = globals.globalScope->lookupType("Backpack"))
-                {
-                    ctv->props["Backpack"] = Luau::makeProperty(backpackType->type);
-                    // TODO: should we duplicate StarterPack into here as well? Is that a reasonable assumption to make?
-                }
-
-                // Player.PlayerGui should contain StarterGui instances
-                if (auto playerGuiType = globals.globalScope->lookupType("PlayerGui"))
-                {
-                    if (auto starterGui = fileResolver.rootSourceNode->findChild("StarterGui"))
-                        addChildrenToCTV(globals, arena, playerGuiType->type, *starterGui);
-                    ctv->props["PlayerGui"] = Luau::makeProperty(playerGuiType->type);
-                }
-
-                // Player.StarterGear should contain StarterPack instances
-                if (auto starterGearType = globals.globalScope->lookupType("StarterGear"))
-                {
-                    if (auto starterPack = fileResolver.rootSourceNode->findChild("StarterPack"))
-                        addChildrenToCTV(globals, arena, starterGearType->type, *starterPack);
-
-                    ctv->props["StarterGear"] = Luau::makeProperty(starterGearType->type);
-                }
-
-                // Player.PlayerScripts should contain StarterPlayerScripts instances
-                if (auto playerScriptsType = globals.globalScope->lookupType("PlayerScripts"))
-                {
-                    if (auto starterPlayer = fileResolver.rootSourceNode->findChild("StarterPlayer"))
-                    {
-                        if (auto starterPlayerScripts = starterPlayer.value()->findChild("StarterPlayerScripts"))
-                        {
-                            addChildrenToCTV(globals, arena, playerScriptsType->type, *starterPlayerScripts);
-                        }
-                    }
-                    ctv->props["PlayerScripts"] = Luau::makeProperty(playerScriptsType->type);
-                }
-            }
-        }
-    }
-
-    // Prepare module scope so that we can dynamically reassign the type of "script" to retrieve instance info
-    frontend.prepareModuleScope = [&frontend, &fileResolver, &arena, expressiveTypes](
-                                      const Luau::ModuleName& name, const Luau::ScopePtr& scope, bool forAutocomplete)
-    {
-        Luau::GlobalTypes& globals = forAutocomplete ? frontend.globalsForAutocomplete : frontend.globals;
-
-        // TODO: we hope to remove these in future!
-        if (!expressiveTypes && !forAutocomplete)
-        {
-            scope->bindings[Luau::AstName("script")] = Luau::Binding{globals.builtinTypes->anyType};
-            scope->bindings[Luau::AstName("workspace")] = Luau::Binding{globals.builtinTypes->anyType};
-            scope->bindings[Luau::AstName("game")] = Luau::Binding{globals.builtinTypes->anyType};
-        }
-
-        if (expressiveTypes || forAutocomplete)
-            if (auto node =
-                    fileResolver.isVirtualPath(name) ? fileResolver.getSourceNodeFromVirtualPath(name) : fileResolver.getSourceNodeFromRealPath(name))
-                scope->bindings[Luau::AstName("script")] = Luau::Binding{getSourcemapType(globals, arena, node.value())};
-    };
-}
-
-// Since in Roblox land, debug is extended to introduce more methods, but the api-docs
-// mark the package name as `@luau` instead of `@roblox`
-static void fixDebugDocumentationSymbol(Luau::TypeId ty, const std::string& libraryName)
-{
-    auto mutableTy = Luau::asMutable(ty);
-    auto newDocumentationSymbol = mutableTy->documentationSymbol.value();
-    replace(newDocumentationSymbol, "@roblox", "@luau");
-    mutableTy->documentationSymbol = newDocumentationSymbol;
-
-    if (auto* ttv = Luau::getMutable<Luau::TableType>(ty))
-    {
-        ttv->name = "typeof(" + libraryName + ")";
-        for (auto& [name, prop] : ttv->props)
-        {
-            newDocumentationSymbol = prop.documentationSymbol.value();
-            replace(newDocumentationSymbol, "@roblox", "@luau");
-            prop.documentationSymbol = newDocumentationSymbol;
-        }
-    }
-}
-
-static auto createMagicFunctionTypeLookup(const std::vector<std::string>& lookupList, const std::string& errorMessagePrefix)
+Luau::MagicFunction createMagicFunctionTypeLookup(const std::vector<std::string>& lookupList, const std::string& errorMessagePrefix)
 {
     return [lookupList, errorMessagePrefix](Luau::TypeChecker& typeChecker, const Luau::ScopePtr& scope, const Luau::AstExprCall& expr,
                const Luau::WithPredicate<Luau::TypePackId>& withPredicate) -> std::optional<Luau::WithPredicate<Luau::TypePackId>>
@@ -481,7 +100,7 @@ static auto createMagicFunctionTypeLookup(const std::vector<std::string>& lookup
     };
 }
 
-std::optional<DefinitionsFileMetadata> parseDefinitionsFileMetadata(const std::string& definitions)
+std::optional<nlohmann::json> parseDefinitionsFileMetadata(const std::string& definitions)
 {
     auto firstLine = getFirstLine(definitions);
     if (Luau::startsWith(firstLine, "--#METADATA#"))
@@ -492,155 +111,11 @@ std::optional<DefinitionsFileMetadata> parseDefinitionsFileMetadata(const std::s
     return std::nullopt;
 }
 
-Luau::LoadDefinitionFileResult registerDefinitions(Luau::Frontend& frontend, Luau::GlobalTypes& globals, const std::string& definitions,
-    bool typeCheckForAutocomplete, std::optional<DefinitionsFileMetadata> metadata)
+Luau::LoadDefinitionFileResult registerDefinitions(
+    Luau::Frontend& frontend, Luau::GlobalTypes& globals, const std::string& definitions, bool typeCheckForAutocomplete)
 {
-    // TODO: packageName shouldn't just be "@roblox"
-    auto loadResult =
-        frontend.loadDefinitionFile(globals, globals.globalScope, definitions, "@roblox", /* captureComments = */ false, typeCheckForAutocomplete);
-    if (!loadResult.success)
-        return loadResult;
-
-    // HACK: Mark "debug" using `@luau` symbol instead
-    if (auto it = globals.globalScope->bindings.find(Luau::AstName("debug")); it != globals.globalScope->bindings.end())
-    {
-        auto newDocumentationSymbol = it->second.documentationSymbol.value();
-        replace(newDocumentationSymbol, "@roblox", "@luau");
-        it->second.documentationSymbol = newDocumentationSymbol;
-        fixDebugDocumentationSymbol(it->second.typeId, "debug");
-    }
-
-    // HACK: Mark "utf8" using `@luau` symbol instead
-    if (auto it = globals.globalScope->bindings.find(Luau::AstName("utf8")); it != globals.globalScope->bindings.end())
-    {
-        auto newDocumentationSymbol = it->second.documentationSymbol.value();
-        replace(newDocumentationSymbol, "@roblox", "@luau");
-        it->second.documentationSymbol = newDocumentationSymbol;
-        fixDebugDocumentationSymbol(it->second.typeId, "utf8");
-    }
-
-    // Extend Instance types
-    if (auto instanceType = globals.globalScope->lookupType("Instance"))
-    {
-        if (auto* ctv = Luau::getMutable<Luau::ClassType>(instanceType->type))
-        {
-            Luau::attachMagicFunction(ctv->props["IsA"].type(), types::magicFunctionInstanceIsA);
-            Luau::attachMagicFunction(ctv->props["FindFirstChildWhichIsA"].type(), types::magicFunctionFindFirstXWhichIsA);
-            Luau::attachMagicFunction(ctv->props["FindFirstChildOfClass"].type(), types::magicFunctionFindFirstXWhichIsA);
-            Luau::attachMagicFunction(ctv->props["FindFirstAncestorWhichIsA"].type(), types::magicFunctionFindFirstXWhichIsA);
-            Luau::attachMagicFunction(ctv->props["FindFirstAncestorOfClass"].type(), types::magicFunctionFindFirstXWhichIsA);
-            Luau::attachMagicFunction(ctv->props["Clone"].type(), types::magicFunctionInstanceClone);
-            Luau::attachMagicFunction(ctv->props["GetPropertyChangedSignal"].type(), magicFunctionGetPropertyChangedSignal);
-
-            // Autocomplete ClassNames for :IsA("") and counterparts
-            Luau::attachTag(ctv->props["IsA"].type(), "ClassNames");
-            Luau::attachTag(ctv->props["FindFirstChildWhichIsA"].type(), "ClassNames");
-            Luau::attachTag(ctv->props["FindFirstChildOfClass"].type(), "ClassNames");
-            Luau::attachTag(ctv->props["FindFirstAncestorWhichIsA"].type(), "ClassNames");
-            Luau::attachTag(ctv->props["FindFirstAncestorOfClass"].type(), "ClassNames");
-
-            // Autocomplete Properties for :GetPropertyChangedSignal("")
-            Luau::attachTag(ctv->props["GetPropertyChangedSignal"].type(), "Properties");
-
-            // Go through all the defined classes and if they are a subclass of Instance then give them the
-            // same metatable identity as Instance so that equality comparison works.
-            // NOTE: This will OVERWRITE any metatables set on these classes!
-            // We assume that all subclasses of instance don't have any metamethods
-            for (auto& [_, ty] : globals.globalScope->exportedTypeBindings)
-            {
-                if (auto* c = Luau::getMutable<Luau::ClassType>(ty.type))
-                {
-                    // Check if the ctv is a subclass of instance
-                    if (Luau::isSubclass(c, ctv))
-                    {
-                        c->metatable = ctv->metatable;
-                    }
-                }
-            }
-        }
-    }
-
-    // Attach onto Instance.new()
-    if (metadata.has_value() && !metadata->CREATABLE_INSTANCES.empty())
-        if (auto instanceGlobal = globals.globalScope->lookup(Luau::AstName("Instance")))
-            if (auto ttv = Luau::get<Luau::TableType>(instanceGlobal.value()))
-                if (auto newFunction = ttv->props.find("new");
-                    newFunction != ttv->props.end() && Luau::get<Luau::FunctionType>(newFunction->second.type()))
-                {
-
-                    Luau::attachTag(newFunction->second.type(), "CreatableInstances");
-                    Luau::attachMagicFunction(
-                        newFunction->second.type(), createMagicFunctionTypeLookup(metadata->CREATABLE_INSTANCES, "Invalid class name"));
-                }
-
-    // Attach onto `game:GetService()`
-    if (metadata.has_value() && !metadata->SERVICES.empty())
-        if (auto serviceProviderType = globals.globalScope->lookupType("ServiceProvider"))
-            if (auto* ctv = Luau::getMutable<Luau::ClassType>(serviceProviderType->type);
-                ctv && Luau::get<Luau::FunctionType>(ctv->props["GetService"].type()))
-            {
-                Luau::attachTag(ctv->props["GetService"].type(), "Services");
-                Luau::attachMagicFunction(ctv->props["GetService"].type(), createMagicFunctionTypeLookup(metadata->SERVICES, "Invalid service name"));
-            }
-
-    // Move Enums over as imported type bindings
-    std::unordered_map<Luau::Name, Luau::TypeFun> enumTypes{};
-    for (auto it = globals.globalScope->exportedTypeBindings.begin(); it != globals.globalScope->exportedTypeBindings.end();)
-    {
-        auto erase = false;
-        auto ty = it->second.type;
-        if (auto* ctv = Luau::getMutable<Luau::ClassType>(ty))
-        {
-            if (Luau::startsWith(ctv->name, "Enum"))
-            {
-                if (ctv->name == "EnumItem")
-                {
-                    Luau::attachMagicFunction(ctv->props["IsA"].type(), types::magicFunctionEnumItemIsA);
-                    Luau::attachTag(ctv->props["IsA"].type(), "Enums");
-                }
-                else if (ctv->name != "Enum" && ctv->name != "Enums")
-                {
-                    // Erase the "Enum" at the start
-                    ctv->name = ctv->name.substr(4);
-
-                    // Move the enum over to the imported types if it is not internal, otherwise rename the type
-                    if (endsWith(ctv->name, "_INTERNAL"))
-                    {
-                        ctv->name.erase(ctv->name.rfind("_INTERNAL"), 9);
-                    }
-                    else
-                    {
-                        enumTypes.emplace(ctv->name, it->second);
-                        // Erase the metatable for the type, so it can be used in comparison
-                    }
-
-                    // Update the documentation symbol
-                    Luau::asMutable(ty)->documentationSymbol = "@roblox/enum/" + ctv->name;
-                    for (auto& [name, prop] : ctv->props)
-                    {
-                        prop.documentationSymbol = "@roblox/enum/" + ctv->name + "." + name;
-                        Luau::attachTag(prop, "EnumItem");
-                    }
-
-                    // Prefix the name (after it has been placed into enumTypes) with "Enum."
-                    ctv->name = "Enum." + ctv->name;
-
-                    erase = true;
-                }
-
-                // Erase the metatable from the type to allow comparison
-                ctv->metatable = std::nullopt;
-            }
-        }
-
-        if (erase)
-            it = globals.globalScope->exportedTypeBindings.erase(it);
-        else
-            ++it;
-    }
-    globals.globalScope->importedTypeBindings.emplace("Enum", enumTypes);
-
-    return loadResult;
+    // TODO: packageName shouldn't just be "@lsp"
+    return frontend.loadDefinitionFile(globals, globals.globalScope, definitions, "@lsp", /* captureComments = */ false, typeCheckForAutocomplete);
 }
 
 using NameOrExpr = std::variant<std::string, Luau::AstExpr*>;
