+++ src/operations/CodeAction.cpp
@@ -69,63 +69,6 @@ lsp::WorkspaceEdit WorkspaceFolder::computeOrganiseRequiresEdit(const lsp::Docum
     return workspaceEdit;
 }
 
-lsp::WorkspaceEdit WorkspaceFolder::computeOrganiseServicesEdit(const lsp::DocumentUri& uri)
-{
-    auto moduleName = fileResolver.getModuleName(uri);
-    auto textDocument = fileResolver.getTextDocument(uri);
-
-    if (!textDocument)
-        throw JsonRpcException(lsp::ErrorCode::RequestFailed, "No managed text document for " + uri.toString());
-
-    frontend.parse(moduleName);
-
-    auto sourceModule = frontend.getSourceModule(moduleName);
-    if (!sourceModule)
-        return {};
-
-    // Find all `local X = game:GetService("Service")`
-    FindImportsVisitor visitor;
-    visitor.visit(sourceModule->root);
-
-    if (visitor.serviceLineMap.empty())
-        return {};
-
-    // Test to see that if all the services are already sorted -> if they are, then just leave alone
-    // to prevent clogging the undo history stack
-    Luau::Location previousServiceLocation{{0, 0}, {0, 0}};
-    bool isSorted = true;
-    for (const auto& [_, stat] : visitor.serviceLineMap)
-    {
-        if (stat->location.begin < previousServiceLocation.begin)
-        {
-            isSorted = false;
-            break;
-        }
-        previousServiceLocation = stat->location;
-    }
-    if (isSorted)
-        return {};
-
-    std::vector<lsp::TextEdit> edits;
-    // We firstly delete all the previous services, as they will be added later
-    edits.reserve(visitor.serviceLineMap.size());
-    for (const auto& [_, stat] : visitor.serviceLineMap)
-        edits.emplace_back(lsp::TextEdit{{{stat->location.begin.line, 0}, {stat->location.begin.line + 1, 0}}, ""});
-
-    // We find the first line to add these services to, and then add them in sorted order
-    lsp::Range insertLocation{{visitor.firstServiceDefinitionLine.value(), 0}, {visitor.firstServiceDefinitionLine.value(), 0}};
-    for (const auto& [serviceName, stat] : visitor.serviceLineMap)
-    {
-        // We need to rewrite the statement as we expected it
-        auto importText = Luau::toString(stat) + "\n";
-        edits.emplace_back(lsp::TextEdit{insertLocation, importText});
-    }
-
-    lsp::WorkspaceEdit workspaceEdit;
-    workspaceEdit.changes.emplace(uri.toString(), edits);
-    return workspaceEdit;
-}
-
 lsp::CodeActionResult WorkspaceFolder::codeAction(const lsp::CodeActionParams& params)
 {
     std::vector<lsp::CodeAction> result;
@@ -133,8 +76,7 @@ lsp::CodeActionResult WorkspaceFolder::codeAction(const lsp::CodeActionParams& p
     auto config = client->getConfiguration(rootUri);
 
     // Add organise imports code action
-    if (params.context.only.empty() || contains(params.context.only, lsp::CodeActionKind::Source) ||
-        contains(params.context.only, lsp::CodeActionKind::SourceOrganizeImports))
+    if (params.context.wants(lsp::CodeActionKind::Source) || params.context.wants(lsp::CodeActionKind::SourceOrganizeImports))
     {
         // Add sort requires code action
         lsp::CodeAction organiseImportsAction;
@@ -147,17 +89,9 @@ lsp::CodeActionResult WorkspaceFolder::codeAction(const lsp::CodeActionParams& p
         //     contains(client->capabilities.textDocument->codeAction->resolveSupport->properties, "edit"))
         organiseImportsAction.edit = computeOrganiseRequiresEdit(params.textDocument.uri);
         result.emplace_back(organiseImportsAction);
-
-        // If in Roblox mode, add a sort services code action
-        if (config.types.roblox)
-        {
-            lsp::CodeAction sortServicesAction;
-            sortServicesAction.title = "Sort services";
-            sortServicesAction.kind = lsp::CodeActionKind::SourceOrganizeImports;
-            sortServicesAction.edit = computeOrganiseServicesEdit(params.textDocument.uri);
-            result.emplace_back(sortServicesAction);
-        }
     }
+
+    platform->handleCodeAction(params, result);
 
     return result;
 }
